# frontend/swagger.json
{"openapi":"3.1.0","info":{"title":"OceanEmbed API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/predict":{"post":{"summary":"Predict","operationId":"predict_predict_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profile":{"get":{"summary":"Profile","operationId":"profile_profile_get","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string","title":"Date"}},{"name":"latitude","in":"query","required":true,"schema":{"type":"number","title":"Latitude"}},{"name":"longitude","in":"query","required":true,"schema":{"type":"number","title":"Longitude"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/surface":{"get":{"summary":"Surface","operationId":"surface_surface_get","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string","title":"Date"}},{"name":"variable","in":"query","required":true,"schema":{"type":"string","title":"Variable"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"summary":"Metrics","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/validation/argo":{"get":{"summary":"Validation Argo","operationId":"validation_argo_validation_argo_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"model_loaded":{"type":"boolean","title":"Model Loaded"},"device":{"type":"string","title":"Device"},"model":{"type":"string","title":"Model"},"depth_count":{"type":"integer","title":"Depth Count"},"grid_resolution_deg":{"type":"number","title":"Grid Resolution Deg"},"date_start":{"type":"string","title":"Date Start"},"date_end":{"type":"string","title":"Date End"}},"type":"object","required":["status","model_loaded","device","model","depth_count","grid_resolution_deg","date_start","date_end"],"title":"HealthResponse"},"Location":{"properties":{"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"}},"type":"object","required":["latitude","longitude"],"title":"Location"},"MatchedGrid":{"properties":{"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"},"distance_km":{"type":"number","title":"Distance Km"}},"type":"object","required":["latitude","longitude","distance_km"],"title":"MatchedGrid"},"PredictRequest":{"properties":{"date":{"type":"string","title":"Date","description":"Date in YYYY-MM-DD format"},"latitude":{"type":"number","maximum":30.0,"minimum":5.0,"title":"Latitude","description":"Latitude between 5.0 and 30.0"},"longitude":{"type":"number","maximum":105.0,"minimum":45.0,"title":"Longitude","description":"Longitude between 45.0 and 105.0"}},"type":"object","required":["date","latitude","longitude"],"title":"PredictRequest"},"PredictResponse":{"properties":{"status":{"type":"string","title":"Status"},"model":{"type":"string","title":"Model"},"date":{"type":"string","title":"Date"},"requested_location":{"$ref":"#/components/schemas/Location"},"matched_grid":{"$ref":"#/components/schemas/MatchedGrid"},"depths_m":{"items":{"type":"number"},"type":"array","title":"Depths M"},"temperature_c":{"items":{"type":"number"},"type":"array","title":"Temperature C"},"temperature_profile":{"additionalProperties":{"type":"number"},"type":"object","title":"Temperature Profile"}},"type":"object","required":["status","model","date","requested_location","matched_grid","depths_m","temperature_c","temperature_profile"],"title":"PredictResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}